clean-up multipliers
